+++ dsp-shared/src/main/scala/dsp/valueobjects/Iri.scala
@@ -8,6 +8,56 @@ package dsp.valueobjects
 import zio.prelude.Validation
 import dsp.errors.BadRequestException
 
+import java.util.UUID
+
+/**
+ * Stores the user ID, i.e. UUID and IRI of the user
+ *
+ * @param uuid the UUID of the user
+ * @param iri the IRI of the user
+ */
+abstract case class UserId private (
+  uuid: UUID,
+  iri: Iri.UserIri
+)
+
+/**
+ * Companion object for UserId. Contains factory methods for creating UserId instances.
+ */
+object UserId {
+
+  /**
+   * Generates a UserId instance with a new (random) UUID and a given IRI which is created from a prefix and the UUID.
+   *
+   * @return a new UserId instance
+   */
+  def fromIri(iri: Iri.UserIri): UserId = {
+    val uuid: UUID = UUID.fromString(iri.value.split("/").last)
+    new UserId(uuid, iri) {}
+  }
+
+  /**
+   * Generates a UserId instance from a given UUID and an IRI which is created from a prefix and the UUID.
+   *
+   * @return a new UserId instance
+   */
+  def fromUuid(uuid: UUID): UserId = {
+    val iri = Iri.UserIri.make("http://rdfh.ch/users/" + uuid.toString).fold(e => throw e.head, v => v)
+    new UserId(uuid, iri) {}
+  }
+
+  /**
+   * Generates a UserId instance with a new (random) UUID and an IRI which is created from a prefix and the UUID.
+   *
+   * @return a new UserId instance
+   */
+  def make(): UserId = {
+    val uuid: UUID = UUID.randomUUID()
+    val iri        = Iri.UserIri.make("http://rdfh.ch/users/" + uuid.toString).fold(e => throw e.head, v => v)
+    new UserId(uuid, iri) {}
+  }
+}
+
 sealed trait Iri
 object Iri {
