+++ dsp-user/core/src/main/scala/dsp/user/domain/UserDomain.scala
@@ -6,93 +6,10 @@
 package dsp.user.domain
 
 import dsp.valueobjects.User._
+import dsp.valueobjects.UserId
 import zio.prelude.Validation
 
 import java.util.UUID
-
-// move this to shared value objects project once we have it
-sealed trait Iri
-object Iri {
-
-  /**
-   * UserIri value object.
-   */
-  sealed abstract case class UserIri private (value: String) extends Iri
-  object UserIri { self =>
-
-    def make(value: String): UserIri = new UserIri(value) {}
-  }
-  // ...
-
-}
-
-/**
- * Stores the user ID, i.e. UUID and IRI of the user
- *
- * @param uuid the UUID of the user
- * @param iri the IRI of the user
- */
-abstract case class UserId private (
-  uuid: UUID,
-  iri: Iri.UserIri
-)
-
-/**
- * Companion object for UserId. Contains factory methods for creating UserId instances.
- */
-object UserId {
-
-  /**
-   * Generates a UserId instance from a given string (either UUID or IRI).
-   *
-   * @param value the string to parse (either UUID or IRI)
-   * @return a new UserId instance
-   */
-  // TODO not sure if we need this
-  // def fromString(value: String): UserId = {
-  //   val uuidPattern = "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$".r
-  //   val iriPattern  = "^http*".r
-
-  //   value match {
-  //     case uuidPattern(value) => new UserId(UUID.fromString(value), Iri.UserIri.make(value)) {}
-  //     case iriPattern(value) =>
-  //       new UserId(UUID.fromString(value.substring(value.lastIndexOf("/") + 1)), Iri.UserIri.make(value)) {}
-  //     //case _                  => ???
-  //   }
-  // }
-
-  /**
-   * Generates a UserId instance with a new (random) UUID and an IRI which is created from a prefix and the UUID.
-   *
-   * @return a new UserId instance
-   */
-  def fromIri(iri: Iri.UserIri): UserId = {
-    val uuid: UUID = UUID.fromString(iri.value.split("/").last)
-    new UserId(uuid, iri) {}
-  }
-
-  /**
-   * Generates a UserId instance with a new (random) UUID and an IRI which is created from a prefix and the UUID.
-   *
-   * @return a new UserId instance
-   */
-  def fromUuid(uuid: UUID): UserId = {
-    val iri: Iri.UserIri = Iri.UserIri.make("http://rdfh.ch/users/" + uuid.toString)
-    new UserId(uuid, iri) {}
-  }
-
-  /**
-   * Generates a UserId instance with a new (random) UUID and an IRI which is created from a prefix and the UUID.
-   *
-   * @return a new UserId instance
-   */
-  // TODO should this return a Validation[Throwable, UserId]
-  def make(): UserId = {
-    val uuid: UUID       = UUID.randomUUID()
-    val iri: Iri.UserIri = Iri.UserIri.make("http://rdfh.ch/users/" + uuid.toString)
-    new UserId(uuid, iri) {}
-  }
-}
 
 /**
  * Represents the user domain object.
